Simplify: Counter for dup check, comprehensions, drop no-op astype
- _check_distinct_authors: manual dup-counting loop -> collections.Counter
- _extract_reasoning: nested loop -> single comprehension
- plots.pairwise_heatmap: nested append loop -> nested comprehension
- pairwise.copeland: drop redundant .astype(int) on bool.sum() results

Behavior-preserving; all tests, ruff, mypy --strict pass.

+++ src/tournament_eval/aggregation/pairwise.py
@@ -205,5 +205,5 @@ def copeland(
 
     beats = above > above.T
     beaten = above < above.T
-    net = beats.sum(axis=1).astype(int) - beaten.sum(axis=1).astype(int)
+    net = beats.sum(axis=1) - beaten.sum(axis=1)
     return {label: float(net[i]) for i, label in enumerate(tally.contestants)}
